chore: move bcrypt and uuid to User model
